API ready

+++ shop/views.py
@@ -1,6 +1,6 @@
 from django.views import generic
 from django.utils import translation
-from django.shortcuts import get_object_or_404
+from django.shortcuts import get_object_or_404 
 
 from rest_framework import status, generics
 from rest_framework.views import APIView
@@ -21,6 +21,7 @@ from .serializers import (
     AdsUserSerializer,
     CategorySerializer, 
     AdsSerializer, 
+    GroupChannelSerializer
 )
 
 def get_query_by_header(self, queryset):
@@ -35,7 +36,7 @@ class BotUserView(generics.ListCreateAPIView):
     serializer_class = BotUserSerializer
 
 
-class ProductUserView(generics.ListCreateAPIView):
+class AdsUserView(generics.ListCreateAPIView):
     queryset = AdsUser.objects.all()
     serializer_class = AdsUserSerializer
 
@@ -76,51 +77,55 @@ class BasicPagination(PageNumberPagination):
     page_size_query_param = 'limit'
 
       
-class AdvertisementView(generics.ListCreateAPIView):
+class AdsListCreateAPIView(generics.ListCreateAPIView):
     serializer_class = AdsSerializer
 
     def get_queryset(self):
         queryset = Advertisement.objects.all().select_related('ads_user', 'group_channel').prefetch_related('categories')
-
         chat_id = self.request.query_params.get('chat_id')
         message_id = self.request.query_params.get('message_id')
-
         if chat_id!=None and message_id!=None:
             queryset = queryset.filter(group_channel__chat_id=chat_id, message_id=message_id)
-            
         return queryset
 
 
     def create(self, request, *args, **kwargs):
-        group_channel, created_group_channel = TelegramGroupChannel.objects.get_or_create(
-            chat_id = request.data['group_channel']['chat_id'],
-            name = request.data['group_channel']['name'],
-            link = request.data['group_channel']['link'],)
+        group_channel = TelegramGroupChannel.objects.filter(
+                        chat_id = request.data.get('group_channel')['chat_id'],
+                        link    = request.data.get('group_channel')['link']).first()
+        if group_channel==None:
+            serializer = GroupChannelSerializer(data=request.data.get('group_channel'))
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-        ads_user, created_ads = AdsUser.objects.get_or_create(
-            user_id = request.data['ads_user']['user_id'], 
-            user_name = request.data['ads_user']['user_name'],
-            user_link = request.data['ads_user']['user_link'],
-            phone_number = request.data['ads_user']['phone_number'],)
-
+        if request.data.get('ads_user'):
+            ads_user = AdsUser.objects.filter(
+                            user_id   = request.data.get('ads_user')['user_id'], 
+                            user_link = request.data.get('ads_user')['user_link']).first()
+            if ads_user==None:
+                ads_user = AdsUser.objects.create(
+                    user_id      = request.data.get('ads_user')['user_id'], 
+                    user_name    = request.data.get('ads_user')['user_name'],
+                    user_link    = request.data.get('ads_user')['user_link'],
+                    phone_number = request.data.get('ads_user')['phone_number'])
         try:
-            advertisement = Advertisement.objects.get(ads_user=ads_user, message_text=request.data['message_text'])
-            if advertisement:
-                advertisement.group_channel = group_channel
-                advertisement.message_id = request.data['message_id']
-                advertisement.datetime = request.data['datetime']
-                advertisement.save()
-                serializer = AdsSerializer(advertisement)
-                return Response(serializer.data, status=status.HTTP_302_FOUND)
-            
+            advertisement = Advertisement.objects.get(ads_user=ads_user, message_text=request.data.get('message_text'))
+            advertisement.group_channel = group_channel
+            advertisement.message_id = request.data.get('message_id')
+            advertisement.datetime = request.data.get('datetime')
+            advertisement.save()
+            serializer = AdsSerializer(advertisement)
+            return Response(serializer.data, status=status.HTTP_302_FOUND)
         except (Advertisement.DoesNotExist) as e:
             advertisement = Advertisement.objects.create(
                 ads_user = ads_user,
                 group_channel = group_channel,
                 message_id = request.data['message_id'],
                 message_text = request.data['message_text'],
-                datetime = request.data['datetime'],
-            )
+                datetime = request.data['datetime'])
+            
             for category in request.data['categories']:
                 advertisement.categories.add(category)
 
@@ -128,43 +133,45 @@ class AdvertisementView(generics.ListCreateAPIView):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-    # def delete(self, request):
-    #     group_id = request.GET.get('group_id')
-    #     message_id = request.GET.get('message_id')
+class AdsRetrieveDestroyAPIView(generics.RetrieveDestroyAPIView):
+    serializer_class = AdsSerializer
+    queryset = Advertisement.objects.all() \
+                .select_related('ads_user', 'group_channel') \
+                .prefetch_related('categories')
+    
+
+class AdvertisementDeleteView(APIView):
+    def delete(self, request):
+        group_channal_id = request.GET.get('group_channal_id')
+        message_id = request.GET.get('message_id')
         
-    #     if (group_id!=None) and (message_id!=None):
-    #         products = get_object_or_404(Product, group__group_id=group_id, message_id=message_id)
-    #         products.delete()
-    #         return Response(status=status.HTTP_204_NO_CONTENT)
-    #     else:
-    #         return Response(data={"error":"group_id and message_id is not given in params!"},status=status.HTTP_400_BAD_REQUEST)
+        if group_channal_id and message_id:
+            products = get_object_or_404(Advertisement, group_channel__chat_id=group_channal_id, message_id=message_id)
+            products.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(data={"error":"group_id and message_id is not given in params!"},status=status.HTTP_400_BAD_REQUEST)
 
 
-class AdvertisementDetailView(APIView):
-    def get(self, request, pk):
-        products = get_object_or_404(Advertisement, pk=pk)
-        product_serializer = AdsSerializer(products)
-        return Response(product_serializer.data)
-
-    def delete(self, request, pk):
-        products = get_object_or_404(Advertisement, pk=pk)
-        products.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
+class AdvertisementPatchView(APIView):
     def patch(self, request, pk, format=None):
         product = get_object_or_404(Advertisement, pk=pk)
 
-        group, created = TelegramGroupChannel.objects.get_or_create(
-            group_id = request.data['group']['group_id'],
-            group_name = request.data['group']['group_name'],
-            group_link = request.data['group']['group_link'],
-        )
+        group_channel = TelegramGroupChannel.objects.filter(
+                        chat_id = request.data.get('group_channel')['chat_id'],
+                        link    = request.data.get('group_channel')['link']).first()
+        if group_channel==None:
+            serializer = GroupChannelSerializer(data=request.data.get('group_channel'))
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            
 
-        product.group = group
-        product.message_id = request.data['message_id']
-        product.message_text = request.data['message_text']
-        product.media_file = request.data['media_file']
-        product.datetime = request.data['datetime']
+        product.group_channel = group_channel
+        product.message_id = request.data.get('message_id')
+        product.message_text = request.data.get('message_text')
+        product.datetime = request.data.get('datetime')
 
         categories = []
         for category_id in request.data.get('categories'):
@@ -175,24 +182,30 @@ class AdvertisementDetailView(APIView):
                 return Response(data={"error":"Category id does not exist"}, status=status.HTTP_404_NOT_FOUND)
             
         product.categories.set(categories)
-
         product.save()
-
         serializer = AdsSerializer(product)
         return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
-
         
 
-class UpdateProductByGroupId(APIView):
-    def patch(self, request, format=None):
-        group_id = request.GET.get('group_id')
+class AdsUpdateByGroupId(APIView):
+    def get(self, request):
+        chat_id = request.GET.get('chat_id')
         message_id = request.GET.get('message_id')
-        products = get_object_or_404(Advertisement, group__group_id=group_id, message_id=message_id)
-        serializer = AdsSerializer(products, data=request.data, partial=True)
+        advertisements = get_object_or_404(Advertisement, group_channel__chat_id=chat_id, message_id=message_id)
+        serializer = AdsSerializer(advertisements)
+        return Response(serializer.data)
+        
+
+    def patch(self, request, format=None):
+        chat_id = request.GET.get('chat_id')
+        message_id = request.GET.get('message_id')
+        advertisements = get_object_or_404(Advertisement, group_channel__chat_id=chat_id, message_id=message_id)
+        serializer = AdsSerializer(advertisements, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 
 #webappbot
